feat: Integration Tests

+++ bank-backend/src/main/java/com/gusta/bank/domain/model/BankAccount.java
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -20,5 +19,5 @@ public class BankAccount {
     @OneToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", referencedColumnName = "id")
     private User user;
-    private BigDecimal balance;
+    private double balance;
 }
