+++ bank-backend/src/main/java/com/gusta/bank/repository/BankAccountRepository.java
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -21,13 +20,13 @@ public interface BankAccountRepository extends JpaRepository<BankAccount, UUID> 
 
     @Modifying
     @Query("UPDATE BankAccount b SET b.balance = b.balance + :value WHERE b.user.email = :email")
-    void sumValueByEmail(@Param("email") String email, @Param("value") BigDecimal value);
+    void sumValueByEmail(@Param("email") String email, @Param("value") double value);
 
     @Modifying
     @Query("UPDATE BankAccount b SET b.balance = b.balance - :value WHERE b.user.email = :email")
-    void subtractValueByEmail(@Param("email") String userEmail, @Param("value") BigDecimal value);
+    void subtractValueByEmail(@Param("email") String userEmail, @Param("value") double value);
 
 
     @Query("select count(b) > 0 from BankAccount b where b.user.email = :email AND b.balance < :value")
-    boolean valueGreaterThanAccountBalance(@Param("email") String userEmail, @Param("value") BigDecimal value);
+    boolean valueGreaterThanAccountBalance(@Param("email") String userEmail, @Param("value") double value);
 }
